refactor dotcom-build-js into a webpack plugin
also merges the JS-specific things from dotcom-build-webpack-config into
the plugin
