+++ packages/dotcom-build-js/README.md
@@ -1,10 +1,8 @@
 # @financial-times/dotcom-build-js
 
-This module extends the [Page Kit CLI build action][cli] (`page-kit build`) with the ability to build JavaScript that makes use of features that are specific to [ft.com]
+This package exports a Webpack plugin to configure it with the ability to build JavaScript that makes use of features that are specific to [ft.com].
 
-[cli]: https://github.com/Financial-Times/dotcom-page-kit/tree/master/packages/dotcom-build-webpack-config#build
 [ft.com]: https://www.ft.com/
-
 
 ## Getting started
 
@@ -33,9 +31,6 @@ Once setup, this plugin will enable you to use the following features within you
 * [Class properties](https://github.com/tc39/proposal-class-public-fields)
 * [Dynamic import syntax](https://developers.google.com/web/updates/2017/11/dynamic-import)
 
-Several [hooks](#hooks) are provided in order to access and modify the configuration.
-
-
 ## Options
 
 | Option                 | Type    | Default      | Description                                                          |
@@ -44,43 +39,3 @@ Several [hooks](#hooks) are provided in order to access and modify the configura
 | `jsxPragmaFrag`        | String  | `"Fragment"` | See https://babeljs.io/docs/en/babel-preset-react#pragmafrag         |
 
 [1]: https://www.npmjs.com/package/babel-plugin-transform-require-default
-
-## Hooks
-
-This plugin exposes the following hooks as extension points. They are available as constants on the exported `hooks` object.
-
-```js
-import { hooks } from '@financial-times/dotcom-build-js'
-```
-
-_Please note: The hooks below are listed in the order they will be executed._
-
-### `BABEL_PRESET_REACT_OPTIONS`
-
-Configuration options for the [@babel/preset-react] plugin.
-
-[@babel/preset-react]: https://babeljs.io/docs/en/babel-preset-react
-
-### `BABEL_PRESET_TYPESCRIPT_OPTIONS`
-
-Configuration options for the [@babel/preset-typescript] plugin.
-
-[@babel/preset-typescript]: https://babeljs.io/docs/en/babel-preset-typescript
-
-### `BABEL_PLUGIN_CLASS_PROPERTIES_OPTIONS`
-
-Configuration options for the [@babel/plugin-proposal-class-properties] plugin.
-
-[@babel/plugin-proposal-class-properties]: https://babeljs.io/docs/en/babel-plugin-proposal-class-properties
-
-### `BABEL_PLUGIN_TRANSFORM_RUNTIME_OPTIONS`
-
-Configuration options for the [@babel/plugin-transform-runtime] plugin.
-
-[@babel/plugin-transform-runtime]: https://babeljs.io/docs/en/babel-plugin-transform-runtime
-
-### `BABEL_PLUGIN_SYNTAX_DYNAMIC_IMPORT_OPTIONS`
-
-Configuration options for the [@babel/plugin-syntax-dynamic-import] plugin.
-
-[@babel/plugin-syntax-dynamic-import]: https://babeljs.io/docs/en/babel-plugin-syntax-dynamic-import
