+++ packages/dotcom-build-js/src/babel.ts
@@ -1,22 +1,29 @@
-import { hooks } from './hooks'
 import { PluginOptions } from './types'
-import { ConfigContext } from '@financial-times/dotcom-build-webpack-config'
 
-export default (options: PluginOptions = {}, context: ConfigContext) => {
-  const presetReactOptions = context.publish(hooks.BABEL_PRESET_REACT_OPTIONS, {
+function getBabelConfig(options: PluginOptions = {}) {
+  const presetEnvOpts = {
+    targets: [
+      'last 2 Chrome versions',
+      'ie 11',
+      'Safari >= 9.1',
+      'ff ESR',
+      'last 2 Edge versions'
+    ],
+    // Exclude transforms that make all code slower
+    // See https://github.com/facebook/create-react-app/pull/5278
+    exclude: ['transform-typeof-symbol']
+  }
+
+  const presetReactOptions = {
     pragma: options.jsxPragma,
     pragmaFrag: options.jsxPragmaFrag
-  })
+  }
 
-  const presetTypescriptOptions = context.publish(hooks.BABEL_PRESET_TYPESCRIPT_OPTIONS, {
+  const presetTypescriptOptions = {
     jsxPragma: options.jsxPragma
-  })
+  }
 
-  const pluginSyntaxDynamicImportOptions = context.publish(hooks.BABEL_PLUGIN_SYNTAX_DYNAMIC_IMPORT_OPTIONS, {})
-
-  const pluginClassPropertiesOptions = context.publish(hooks.BABEL_PLUGIN_CLASS_PROPERTIES_OPTIONS, {})
-
-  const pluginTransformRuntimeOptions = context.publish(hooks.BABEL_PLUGIN_TRANSFORM_RUNTIME_OPTIONS, {
+  const pluginTransformRuntimeOptions = {
     // You might think we'd want to abstract the helper functions so they can be reused but doing so
     // means we generate unstable hashes because the generated helper modules are at the bottom of
     // the dependency tree but their contents depends on the features each app uses. Inlining them
@@ -24,10 +31,15 @@ export default (options: PluginOptions = {}, context: ConfigContext) => {
     // them down and actually result in fewer function calls overall!
     // <https://github.com/Financial-Times/dotcom-page-kit/issues/576>
     helpers: false
-  })
+  }
 
   const config = {
+    // By default Babel assumes all source code is ESM so force it to check for CJS
+    sourceType: 'unambiguous',
+    babelrc: true,
+    cacheDirectory: true,
     presets: [
+      [require.resolve('@babel/preset-env'), presetEnvOpts],
       [require.resolve('@babel/preset-react'), presetReactOptions],
       // This only enables the parsing of TypeScript, it does not check types
       [require.resolve('@babel/preset-typescript'), presetTypescriptOptions]
@@ -35,12 +47,25 @@ export default (options: PluginOptions = {}, context: ConfigContext) => {
     plugins: [
       // This is required by @babel/preset-typescript
       // https://github.com/tc39/proposal-class-fields
-      [require.resolve('@babel/plugin-proposal-class-properties'), pluginClassPropertiesOptions],
+      [require.resolve('@babel/plugin-proposal-class-properties')],
       // This enables Babel's built-in 'dynamicImport' flag which defines import() function usage
-      [require.resolve('@babel/plugin-syntax-dynamic-import'), pluginSyntaxDynamicImportOptions],
+      [require.resolve('@babel/plugin-syntax-dynamic-import')],
       [require.resolve('@babel/plugin-transform-runtime'), pluginTransformRuntimeOptions]
-    ] as any[]
+    ]
   }
 
   return config
+}
+
+export default function getBabelRule(userOptions: PluginOptions) {
+  return {
+     test: [/\.(js|jsx|mjs|ts|tsx)$/],
+     // NOTE: Do not exclude bower_components or node_modules directories
+     // https://github.com/Financial-Times/dotcom-page-kit/issues/366
+     exclude: [],
+     use: {
+       loader: require.resolve('babel-loader'),
+       options: getBabelConfig(userOptions)
+     }
+   }
 }
