--- packages/dotcom-build-js/src/hooks.ts
@@ -1,7 +1,0 @@
-export enum hooks {
-  BABEL_PRESET_REACT_OPTIONS = 'babelConfig::preset::react::options',
-  BABEL_PRESET_TYPESCRIPT_OPTIONS = 'babelConfig::preset::typescript::options',
-  BABEL_PLUGIN_CLASS_PROPERTIES_OPTIONS = 'babelConfig::plugin::proposalClassProperties::options',
-  BABEL_PLUGIN_SYNTAX_DYNAMIC_IMPORT_OPTIONS = 'babelConfig::plugin::syntaxDynamicImport::options',
-  BABEL_PLUGIN_TRANSFORM_RUNTIME_OPTIONS = 'babelConfig::plugin::transformRuntime::options'
-}
