+++ packages/dotcom-build-js/src/index.ts
@@ -1,5 +1,4 @@
 import { plugin } from './plugin'
 
-export { hooks } from './hooks'
 export { plugin }
 export default plugin
