+++ packages/dotcom-build-js/src/plugin.ts
@@ -1,6 +1,6 @@
-import babelPreset from './babel'
-import { HandlerArgs, hooks } from '@financial-times/dotcom-build-webpack-config'
+import getBabelRule from './babel'
 import { PluginOptions } from './types'
+import type webpack from 'webpack'
 
 const defaultOptions: PluginOptions = {
   jsxPragma: 'h',
@@ -10,22 +10,13 @@ const defaultOptions: PluginOptions = {
 export function plugin(userOptions: PluginOptions = {}) {
   const options = { ...defaultOptions, ...userOptions }
 
-  return ({ on }) => {
-    on(hooks.BABEL_CONFIG, addBabelPreset)
-    on(hooks.WEBPACK_JS_RULE, amendWebpackConfigScriptsRule)
-    on(hooks.WEBPACK_CONFIG, addTypeScriptFileTypesToResolvers)
+  return {
+    apply(compiler: webpack.Compiler) {
+      compiler.options.resolve = {
+        extensions: ['.js', '.jsx', '.mjs', '.json', '.ts', '.tsx']
   }
 
-  function addTypeScriptFileTypesToResolvers({ resource: webpackConfig }: HandlerArgs) {
-    webpackConfig.resolve.extensions.push('.ts', '.tsx')
+      compiler.options.module.rules.push(getBabelRule(options))
   }
-
-  function amendWebpackConfigScriptsRule({ resource: scriptsRule }) {
-    // Replace default JS test with a RegExp including TypeScript file extensions
-    scriptsRule.test.push(/\.(ts|tsx)$/)
-  }
-
-  function addBabelPreset({ context, resource: babelConfig }: HandlerArgs) {
-    babelConfig.presets.push(babelPreset(options, context))
   }
 }
